+++ CPP_Graphics/src/planets/bodies_simulation.h
@@ -4,5 +4,5 @@
 class BodiesSimulation
 {
 public:
-	static void standard(Planet *planets, unsigned int planetCount, float deltaTime);
+	static void standard(Planet *planets, unsigned int planetCount, double deltaTime);
 };
